LayoutTests/fast/scrolling/mac/j-shaped-scroll-rubberband-expected.txt

PASS window.scrollY > 0 is true
